+++ src/components/Dropdown/DropdownElements/DropdownExpandEmoji.jsx
